finishing new yt-mp3 script

+++ bin/yt-mp3.py
@@ -6,6 +6,7 @@ import subprocess
 import re
 import os
 import shutil
+import sys
 from pprint import pprint
 
 mypwd = os.getcwd()
@@ -13,25 +14,24 @@ tempdir = tempfile.mkdtemp()
 os.chdir(tempdir)
 
 #TODO change to variable
-url="http://youtu.be/Xe1dga2K9O4"
+url=sys.argv[1]
 
-match = re.search('.*/([^/&]*)', url)
-vid=''
-if match:
-	vid=match.group(1)
-	print "downloading " + vid
-else:
-	raise Exception('could not get id from url')
-
-subprocess.call(["youtube-dl", "--id", "--write-thumbnail", "--write-info-json", "-x", "--audio-format", "mp3", "--audio-quality", "128K", url])
-
-json_data = open(vid+".info.json")
-metadata = json.load(json_data)
+json_data = subprocess.check_output(["youtube-dl","-j",url])
+metadata = json.loads(json_data)
 title=metadata[u'title']
 artist=metadata[u'uploader']
+vid=metadata[u'id']
 print title
 print artist
-subprocess.call(["lame", "--ti", vid + ".jpg", "--tt", title, "--ta", artist, vid + ".mp3"])
+print vid
+
+if subprocess.call(["youtube-dl", "--id", "--write-thumbnail", "-x", "--audio-format", "mp3", "--audio-quality", "128K", url]) != 0:
+	raise Exception('could not download')
+
+if subprocess.call(["lame", "--ti", vid + ".jpg", "--tt", title, "--ta", artist, vid + ".mp3"]) != 0:
+	raise Exception('could not tag')
+
+os.rename(vid + '.mp3', '/home/rpetti/Music/'+title+'.mp3')
 
 os.chdir(mypwd)
 shutil.rmtree(tempdir)
